Sprint-12 controller tests removed, service tests integrated with databases
