--- src/test/java/ru/yandex/practicum/filmorate/controller/FilmControllerTest.java
@@ -1,16 +1,0 @@
-package ru.yandex.practicum.filmorate.controller;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-public class FilmControllerTest {
-    private FilmController controller;
-
-    @BeforeEach
-    public void testInit() {
-    }
-
-    @Test
-    public void testGetFilmByInvalidIdThrowsNotFoundException() {
-    }
-}
